[CI] switch to pytest to test the package in conda-build (#980)
* [CI] switch to pytest to test the package in conda-build

Background is that nosetests is no longer maintained (>4 years).
Pytest is modern and maintained with tons of available plugins.

Enabled support for CircleCI and CodeCov (coverage service)
Fixed some minor testing issues.

Fixes #743

+++ _base/datasource.py
@@ -219,8 +219,8 @@ class DataSource(Iterable, TrajectoryRandomAccessible):
         skip
         """
         if chunksize != 0:
-            chunks = sum((ceil(l / float(chunksize))
-                          for l in self.trajectory_lengths(stride=stride, skip=skip)))
+            chunks = int(sum((ceil(l / float(chunksize))
+                          for l in self.trajectory_lengths(stride=stride, skip=skip))))
         else:
             chunks = self.number_of_trajectories(stride)
         return chunks
